Changed ticket index controller to send all tickets to currentLogged in user if they are either in ROLE_ADMIN or ROLE_TECH.
We will filter these on the view based on user role and workgroup id.

Techs and admins might have the ability to filter by workgroup instead of having to click on a different view.

If user is not a tech or admin, it only displays their ticketts based on their username=ticket.email.