Replies now notify and add history to ticket
Set SESSION['ticket'] to keep track of the ticket information when going to the reply controller.

Added history item to ticket when a reply is made.

Added email to client when a reply is made by ROLE_TECH or ROLE_ADMIN

Added email to assigned tech when a reply is made by ROLE_USER (patron)

Changed the way I get user infromation from getAuthenticatedUser which is deprecated and throwing exceptions to user-springSecurityService.currentUser.  Need to change this with other controllers.