Changed Close Ticket button to require Reply
CLosed ticket redirects to special replyCloseTicket controller action which requires a reply to then close the ticket.

This controller action saves the reply then redirects to the ticket controller close action.

The close action closes the ticket, emails the correct person, and then redirects back to the ticket index.

Also changed the security on closing and opening tickets to allow regular users to do this.