Ticket no longer belongs to workgroup
User no longer belongs to ticket

User no longer belongs to workgroup

Bug Fix: Admin could not delete a ticket if it had been assigned to a technician or workgroup. Changed domain class so that tickets no longer belong to anything so that they can be deleted.

Bug Fix: Users could not be deleted if assigned to a workgroup or ticket.  Changed domain class so that tickets no longer belong to anything so that they can be deleted.

Bug Fix: Users could not be deleted if they had roles assigned to them.  Added try/catch blocks in the delete action in the UserController to remove patron, tech, or admin roles if applicable and then delete the user.

Bug Fix: User password field was showing plain text when entering password.  Changed to password field.