Added bean instance for author in notes and replies instead of link in the show.gsp
Created a link for changing password (admins only) in the show.gsp for users instead of the encrypted password.

Changed userRole label to User Role in the show.gsp for users.

Added a section on the view for phone search and added a query in index action of ticket controller.

Added a reset action in the ticket controller and button in the ticket control section of the index.gsp to reset the view.

Added a unassigned action to show unassigned tickets only in the ticket controller and added a button for this in the ticket control section of index.gsp

Created a seperate form called formCLoseReply and added that to the createCloseReply.gsp render so that it would not show the wrong message for just creating a reply to a ticket.